+++ app/views/content_items/signin/choose-sign-in.html.erb
@@ -34,7 +34,7 @@
   </div>
 <% end %>
 
-<%= form_tag({controller: 'content_items', action: 'sign_in_options'}, { method: :post, class: 'js-radio-button-track-form govuk-u-mb-6' }) do %>
+<%= form_tag({controller: 'self_assessment_signin', action: 'sign_in_options'}, { method: :post, class: 'js-radio-button-track-form govuk-u-mb-6' }) do %>
   <fieldset class="govuk-u-mb-5 govuk-c-fieldset" id="radio-stacked">
     <legend class="govuk-c-fieldset__legend">
       <h1 class="govuk-u-bold-48 govuk-u-mb-6 app-u-mb-4--small">Choose how to sign in</h1>
